logout, login, routes
Change code:
- delete comments
- delete file contacts-db
- find user by token (no by email)
- reunir methods with routes
