+++ auth/loginHandler.js
@@ -21,22 +21,6 @@ const loginHandler = async (email, incomingPassword) => {
     // eslint-disable-next-line no-throw-literal
     throw { code: 401, msg: "Invalid credentials" };
   }
-
-  // if (!contact) {
-  //   // eslint-disable-next-line no-throw-literal
-  //   throw { code: 404, msg: "User not found" };
-  // }
-
-  // const contactPassword = contact.password;
-
-  // const result = bcrypt.compareSync(incomingPassword, contactPassword);
-
-  // if (result) {
-  //   return issueToken(contact);
-  // } else {
-  //   // eslint-disable-next-line no-throw-literal
-  //   throw { code: 401, msg: "Invalid credentials" };
-  // }
 };
 
 module.exports = loginHandler;
