+++ routes/api/passportStrategy.js
@@ -1,0 +1,27 @@
+const passport = require("passport");
+const passportJWT = require("passport-jwt");
+require("dotenv").config();
+
+const { User } = require("../../models/user");
+
+const secret = process.env.JWT_SECRET;
+
+const ExtractJWT = passportJWT.ExtractJwt;
+const Strategy = passportJWT.Strategy;
+const params = {
+  secretOrKey: secret,
+  jwtFromRequest: ExtractJWT.fromAuthHeaderAsBearerToken(),
+};
+
+passport.use(
+  new Strategy(params, function (payload, done) {
+    User.find({ _id: payload.id })
+      .then(([user]) => {
+        if (!user) {
+          return done(new Error("User not found"));
+        }
+        return done(null, user);
+      })
+      .catch((err) => done(err));
+  })
+);
